analyzer/requirements.txt
pymongo
pandas
python-dotenv
flask
<<<<<<< Updated upstream
gunicorn
=======
gunicorn
pytest
tiktoken
tokonomics
typing_extensions
flask-cors
>>>>>>> Stashed changes
